Move FiveWay, TFT and OLED under SCREEN
