+++ src/targets/axis_2400.ini
@@ -8,9 +8,7 @@ build_flags =
 	${common_env_data.build_flags_tx}
 	-include target/AXIS_THOR_2400_TX.h
 	-D TARGET_AXIS_THOR_2400_TX=1
-	-D USER_SETUP_LOADED=1
-	-D CONFIG_DISABLE_HAL_LOCKS=1
-	-include lib/TFTSCREEN/TFT_eSPI_User_Setup.h
+	-include lib/SCREEN/TFT/TFT_eSPI_User_Setup.h
 	-D VTABLES_IN_FLASH=1
 	-O2
 monitor_speed = 420000
@@ -18,8 +16,6 @@ src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.lib_deps}
     bodmer/TFT_eSPI
-
-
 
 [env:AXIS_THOR_2400_TX_via_WIFI]
 extends = env:AXIS_THOR_2400_TX_via_UART
